+++ app/Http/Controllers/ColisController.php
@@ -3,10 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Client;
+use App\Mail\ColisMail;
+use App\Mail\Colis2Mail;
 use App\Colis;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Tarif;
+use App\User;
+use App\AccountType;
 use Carbon\Carbon;
 use Illuminate\Support\Facades\Mail;
 use Barryvdh\DomPDF\Facade as PDF;
@@ -20,12 +24,29 @@ class ColisController extends Controller
      */
     public function index()
     {
-        $colis = Colis::where("country", auth()->user()->country)->get();
+        // $colis = Colis::where("country", auth()->user()->country)->get();
+        // $colis = Colis::with('user')->get();
+        if(auth()->user()->accountType->code != "ADM" ){
+        $colis = Colis::with('user')
+              ->where("who", auth()->user()->id)
+              ->get();
+        return view("colis.index", compact("colis"));
+        }
+        $colis = Colis::all();
         return view("colis.index", compact("colis"));
     }
 
+    public function tracer()
+    {
+        $colis = Colis::where("user_id", auth()->user()->id)
+              ->get();
+            //   $id =  auth()->user()->id;
+            //   dd($colis);
+        return view("colis.tracer", compact("colis"));
+    }
+
     public function withdrawal(){
-        $colis = Colis::where("country", "!=", auth()->user()->country)->get();
+        $colis = Colis::where("country", "=", auth()->user()->accountType->id)->get();
         return view("colis.withdrawal", compact("colis"));
     }
 
@@ -35,14 +56,19 @@ class ColisController extends Controller
             "message" => "Vous ne pouvez pas retirer un colis en attente de paiement."
         ]);
         $month = Carbon::now()->month;
+        $formattedDate = Carbon::now()->format('d/m/y');
         $colis->statut = "Retiré";
+        $details = [
+            'title' => 'Votre colis a été livré avec succès',
+            'body' => 'Nous somme ravi d avoir traité avec vous, nous espérons vous revoir bientot',
+            'colis1' => $colis->user->first_name,
+            'colis2' => $colis->user->last_name,
+        ];
         $colis->save();
-        /*Mail::raw('Bonjour monsieur, votre colis est arrivé à destination.', function ($message) use($colis) {
-            $message->from('support@zedlogistics.com', 'Support');
-            $message->to($colis->client->email, $colis->client->firstName." ".$colis->client->lastName);
-        });*/
-        $pdf = PDF::loadView('colis.withdrawal_invoice', ["colis" => $colis, "month" => $month]);
-        return $pdf->stream('colis.pdf');
+        Mail::to($colis->user->email)->send(new Colis2Mail($details));
+        
+        return view('colis.withdrawal_invoice_double', ["colis" => $colis, "formattedDate" => $formattedDate]);
+        
     }
 
     /**
@@ -52,9 +78,11 @@ class ColisController extends Controller
      */
     public function create()
     {
-        $clients = Client::all();
+        // $clients = User::all();
+        $clients = User::where('account_id', 4)->get();
         $tarifs = Tarif::all();
-        return view("colis.create", compact("clients", "tarifs"));
+        $accountTypes = AccountType::all();
+        return view("colis.create", compact("clients", "tarifs", "accountTypes"));
     }
 
     /**
@@ -66,21 +94,34 @@ class ColisController extends Controller
     public function store(Request $request)
     {
         $this->validate($request, [
-            "client_id" => "required",
+            "user_id" => "required",
             "receiver_id" => "required",
         ]);
 
-        if($request->client_id == $request->receiver_id) return back()->withErrors([
+        if($request->user_id == $request->receiver_id) return back()->withErrors([
             "message" => "Le client ne peut s'envoyer de colis à lui même."
         ]);
 
         $colis = new Colis();
         $colis->country = auth()->user()->country;
         $colis->date_entree = Carbon::today();
+        $colis->longueur = $request->longueur;
+        $colis->hauteur = $request->hauteur;
+        $colis->poids = $request->poids;
+        $colis->largeur = $request->largeur;
+        $colis->who = auth()->user()->id;
+        //dd($colis->valeur_euro);
         $colis->fill($request->all());
+        $colis->valeur_euro = (($colis->longueur * $colis->hauteur * $colis->largeur) / 5000) * 30;
+        $details = [
+            'title' => 'Votre colis est en cours de traitement',
+            'body' => 'Nous somme ravi d avoir traité avec vous, nous espérons vous revoir bientot',
+            'colis' => $colis->date_entree
+        ];
 
         try{
             $colis->save();
+
             return back()->withSuccess("Le colis a enregistré avec succès.");
 
         }catch(Exception $e){
@@ -99,21 +140,45 @@ class ColisController extends Controller
      */
     public function show($id)
     {
-        $clients = Client::all();
-        $colis = Colis::with("client")->findOrFail($id);
+        $clients = User::all();
+        $colis = Colis::with("user")->findOrFail($id);
         $tarifs = Tarif::all();
         return view("colis.show", compact("colis", "clients", "tarifs"));
     }
 
+
     public function send($id){
         $colis = Colis::findOrFail($id);
         $colis->statut = "Envoyé";
+
+        $details = [
+            'title' => 'Votre colis est en cours de traitement',
+            'body' => 'Nous somme ravi d avoir traité avec vous, nous espérons vous revoir bientot',
+            'colis1' => $colis->receiver->first_name,
+            'colis2' => $colis->receiver->last_name,
+            'colis11' => $colis->user->first_name,
+            'colis22' => $colis->user->last_name,
+            'colis5' => $colis->user->address,
+            'colis6' => $colis->user->country,
+            'colis3' => $colis->nom,
+            'colis4' => $colis->contenance,
+            'colis7' => $colis->receiver->login,
+            'colis8' => $colis->receiver->password_eph,
+            'colis9' => $colis->date_arrivee,
+        ];
         $colis->save();
+        // Send email
+        Mail::to($colis->receiver->email)->send(new ColisMail($details));
 
-        $month = Carbon::now()->month;
+        $colis->receiver->password_eph = "";
 
-        $pdf = PDF::loadView('colis.deposit_invoice', ["colis" => $colis, "month" => $month]);
-        return $pdf->stream('colis.pdf');
+        $month = Carbon::now();
+        $formattedDate = Carbon::now()->format('d/m/y');
+
+        return view('colis.deposit_invoice_double', ["colis" => $colis, "formattedDate" => $formattedDate]);
+
+        // $pdf = PDF::loadView('colis.deposit_invoice', ["colis" => $colis, "month" => $month]);
+        // return $pdf->stream('colis.pdf');
         //return back()->withSuccess("Le colis a bien été envoyé.");
     }
 
@@ -126,7 +191,7 @@ class ColisController extends Controller
     public function edit($id)
     {
         $colis = Colis::findOrFail($id);
-        $clients = Client::all();
+        $clients = User::all();
         $tarifs = Tarif::all();
 
         return view("colis.edit", compact("colis", "clients", "tarifs"));
@@ -172,6 +237,12 @@ class ColisController extends Controller
     public function destroy($id)
     {
         $colis = Colis::findOrFail($id);
+
+        if($colis->count() > 0){
+            return redirect()->back()->withErrors([
+                "message" => "Le colis ne peut être supprimé car possède un incident."
+            ]);
+        }
         $colis->delete();
 
         return back()->withSuccess("Le colis a été supprimé avec succès.");
